+++ docs/nav/development/release-notes.md
@@ -1,5 +1,16 @@
 # Release Notes
 
+## v0.3.0
+
+- [📦 PyPI - Build 0.3.0](https://github.com/FernandoCelmer/dotflow/releases/tag/v0.3.0)
+- [📌 Improvement of the API to start the workflow task](https://github.com/FernandoCelmer/dotflow/issues/10)
+
+## v0.2.0
+
+- [📦 PyPI - Build 0.2.0](https://github.com/FernandoCelmer/dotflow/releases/tag/v0.2.0)
+- [📌 Create documentation](https://github.com/FernandoCelmer/dotflow/issues/6)
+- [📌 Import of the context class](https://github.com/FernandoCelmer/dotflow/issues/9)
+
 ## v0.1.0
 
 - [📦 PyPI - Build 0.1.0](https://github.com/FernandoCelmer/dotflow/releases/tag/v0.1.0)